Update WatchdogNRF.cpp use 64 bit math in reload value calc
The nature of the NRF watch dog requires a multiply by 32768 to calculate the reload register value. If using a longer timeout like 300000 ms (5 minutes) it will overflow and not work as intended. This temporarily converts the multiply & divide to a 64 bit operation before setting the register. So longer a timeout can be used.